Split mixed runtime C files into runtime and stdlib halves
Divide io.c, sys.c, gc_ctrl.c and obj.c, each of which mixed C
implementations of library functions with runtime support code.  The
library halves move to stdlib/prims/<name>_prims.c; the runtime
halves keep their files.  The prims halves need distinct basenames
because both halves are archived into the same runtime libraries,
and ar drops archive members with duplicate names.

- io.c: the caml_ml_* channel primitives move; the buffered-I/O
  machinery and the channel custom operations stay.  The runtime
  newly exports caml_channel_register, caml_flush_noexc and
  caml_channel_check_pending (previously a static) for use by the
  primitives.
- sys.c: the Sys primitives move; error raising, caml_do_exit and
  caml_sys_init stay.  The static main_argv becomes the exported
  caml_main_argv.
- gc_ctrl.c: the Gc primitives move; caml_init_gc, the GC tweaks
  machinery and OCAMLRUNPARAM reflection stay.
- obj.c: the Obj reflection and CamlinternalOO primitives move;
  atomic tag reads, lazy and forwarding support and the let rec
  dummies stay.
