+++ runtime/gc_ctrl.c
@@ -66,111 +66,12 @@ extern atomic_uintnat caml_custom_minor_ratio; /* see custom.c */
 extern atomic_uintnat caml_custom_minor_max_bsz; /* see custom.c */
 extern uintnat caml_minor_heap_max_wsz; /* see domain.c */
 
-CAMLprim value caml_gc_quick_stat(value v)
-{
-  CAMLparam0 ();   /* v is ignored */
-  CAMLlocal1 (res);
-
-  /* get a copy of these before allocating anything... */
-  intnat majcoll, mincoll, compactions;
-  struct gc_stats s;
-  caml_compute_gc_stats(&s);
-  majcoll = caml_major_cycles_completed;
-  mincoll = atomic_load(&caml_minor_collections_count);
-  compactions = atomic_load(&caml_compactions_count);
-
-  res = caml_alloc_tuple (17);
-  Store_field (res, 0, caml_copy_double ((double)s.alloc_stats.minor_words));
-  Store_field (res, 1, caml_copy_double ((double)s.alloc_stats.promoted_words));
-  Store_field (res, 2, caml_copy_double ((double)s.alloc_stats.major_words));
-  Store_field (res, 3, Val_long (mincoll));
-  Store_field (res, 4, Val_long (majcoll));
-  Store_field (res, 5, Val_long (
-    s.global_stats.chunk_words + s.heap_stats.large_words));
-  Store_field (res, 6, Val_long (s.global_stats.chunks));
-  Store_field (res, 7, Val_long (
-    s.heap_stats.pool_live_words + s.heap_stats.large_words));
-  Store_field (res, 8, Val_long (
-    s.heap_stats.pool_live_blocks + s.heap_stats.large_blocks));
-  Store_field (res, 9, Val_long (
-    s.global_stats.chunk_words - s.heap_stats.pool_live_words
-    - s.heap_stats.pool_frag_words));
-  Store_field (res, 10, Val_long (0)); /* free_blocks */
-  Store_field (res, 11, Val_long (0)); /* largest_free */
-  Store_field (res, 12, Val_long (s.heap_stats.pool_frag_words));
-  Store_field (res, 13, Val_long (compactions));
-  Store_field (res, 14, Val_long (
-    s.heap_stats.pool_max_words + s.heap_stats.large_max_words));
-  Store_field (res, 15, Val_long (0)); /* stack_size */
-  Store_field (res, 16, Val_long (s.alloc_stats.forced_major_collections));
-  CAMLreturn (res);
-}
-
-double caml_gc_minor_words_unboxed (void)
-{
-  return (double)caml_minor_words_allocated();
-}
-
-CAMLprim value caml_gc_minor_words(value v)
-{
-  CAMLparam0 ();   /* v is ignored */
-  CAMLreturn(caml_copy_double(caml_gc_minor_words_unboxed()));
-}
-
-CAMLprim value caml_gc_counters(value v)
-{
-  CAMLparam0 (); /* v is ignored */
-  CAMLlocal4 (minwords_, prowords_, majwords_, res);
-
-  /* get a copy of these before allocating anything... */
-  double minwords = caml_gc_minor_words_unboxed();
-  double prowords = (double)Caml_state->stat_promoted_words;
-  double majwords = Caml_state->stat_major_words +
-                    (double) Caml_state->allocated_words;
-
-  minwords_ = caml_copy_double(minwords);
-  prowords_ = caml_copy_double(prowords);
-  majwords_ = caml_copy_double(majwords);
-  res = caml_alloc_3(0, minwords_, prowords_, majwords_);
-  CAMLreturn(res);
-}
-
-CAMLprim value caml_gc_get(value v)
-{
-  CAMLparam0 ();   /* v is ignored */
-  CAMLlocal1 (res);
-
-  res = caml_alloc_tuple (11);
-  Store_field (res, 0, Val_long (Caml_state->minor_heap_wsz));          /* s */
-  Store_field (res, 1,
-    Val_long (atomic_load_relaxed(&caml_major_heap_increment)));        /* i */
-  Store_field (res, 2,
-    Val_long (atomic_load_relaxed(&caml_percent_free)));                /* o */
-  Store_field (res, 3, Val_long (atomic_load_relaxed(&caml_verb_gc)));  /* v */
-  Store_field (res, 4,
-    Val_long (atomic_load_relaxed(&caml_max_percent_free)));
-  Store_field (res, 5, Val_long (caml_max_stack_wsize));                /* l */
-  Store_field (res, 6, Val_long (0));
-  Store_field (res, 7, Val_long (0));
-  Store_field (res, 8,
-    Val_long (atomic_load_relaxed(&caml_custom_major_ratio)));          /* M */
-  Store_field (res, 9,
-    Val_long (atomic_load_relaxed(&caml_custom_minor_ratio)));          /* m */
-  Store_field (res, 10,
-    Val_long (atomic_load_relaxed(&caml_custom_minor_max_bsz)));        /* n */
-  CAMLreturn (res);
-}
-
 #define Max(x,y) ((x) < (y) ? (y) : (x))
 
+/* Kept in sync with the twin definitions in stdlib/prims/gc_ctrl_prims.c. */
 static uintnat norm_pfree (uintnat p)
 {
   return Max (p, 1);
-}
-
-static uintnat norm_pmax (uintnat p)
-{
-  return p;
 }
 
 static uintnat norm_custom_maj (uintnat p)
@@ -181,202 +82,6 @@ static uintnat norm_custom_maj (uintnat p)
 static uintnat norm_custom_min (uintnat p)
 {
   return Max (p, 1);
-}
-
-CAMLprim value caml_gc_set(value v)
-{
-  uintnat newminwsz = caml_norm_minor_heap_size (Long_val (Field (v, 0)));
-  uintnat newheapincr = Long_val (Field (v, 1));
-  uintnat newpf = norm_pfree (Long_val (Field (v, 2)));
-  uintnat new_verb_gc = Long_val (Field (v, 3));
-  uintnat newpm = norm_pmax (Long_val (Field (v, 4)));
-  uintnat new_max_stack_size = Long_val (Field (v, 5));
-  /* ignore fields 6 (allocation policy) and 7 (window size) */
-  uintnat new_custom_maj = norm_custom_maj (Long_val (Field (v, 8)));
-  uintnat new_custom_min = norm_custom_min (Long_val (Field (v, 9)));
-  uintnat new_custom_sz = Long_val (Field (v, 10));
-
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_SET);
-
-  if (newheapincr != atomic_load_relaxed(&caml_major_heap_increment)) {
-    atomic_store_relaxed(&caml_major_heap_increment, newheapincr);
-    if (newheapincr > 1000) {
-      CAML_GC_MESSAGE(PARAMS, "New heap increment size: %"
-                      ARCH_INTNAT_PRINTF_FORMAT "uk words\n",
-                      newheapincr/1024);
-    } else {
-      CAML_GC_MESSAGE(PARAMS, "New heap increment size: %"
-                      ARCH_INTNAT_PRINTF_FORMAT "u%%\n",
-                      newheapincr);
-    }
-  }
-
-  caml_change_max_stack_size (new_max_stack_size);
-
-  if (newpf != atomic_load_relaxed(&caml_percent_free)){
-    atomic_store_relaxed(&caml_percent_free, newpf);
-    CAML_GC_MESSAGE(PARAMS,
-                    "New space overhead: %" ARCH_INTNAT_PRINTF_FORMAT "u%%\n",
-                    newpf);
-  }
-  if (newpm != atomic_load_relaxed(&caml_max_percent_free)) {
-    atomic_store_relaxed(&caml_max_percent_free, newpm);
-    CAML_GC_MESSAGE(PARAMS, "New max space overhead: %"
-                    ARCH_INTNAT_PRINTF_FORMAT "u%%\n", newpm);
-  }
-
-  atomic_store_relaxed(&caml_verb_gc, new_verb_gc);
-
-  /* These fields were added in 4.08.0. */
-  if (Wosize_val (v) >= 11){
-    if (new_custom_maj != atomic_load_relaxed(&caml_custom_major_ratio)){
-      atomic_store_relaxed(&caml_custom_major_ratio, new_custom_maj);
-      CAML_GC_MESSAGE(PARAMS, "New custom major ratio: %"
-                      ARCH_INTNAT_PRINTF_FORMAT "u%%\n",
-                      new_custom_maj);
-    }
-    if (new_custom_min != atomic_load_relaxed(&caml_custom_minor_ratio)){
-      atomic_store_relaxed(&caml_custom_minor_ratio, new_custom_min);
-      CAML_GC_MESSAGE(PARAMS, "New custom minor ratio: %"
-                      ARCH_INTNAT_PRINTF_FORMAT "u%%\n",
-                      new_custom_min);
-    }
-    if (new_custom_sz != atomic_load_relaxed(&caml_custom_minor_max_bsz)){
-      atomic_store_relaxed(&caml_custom_minor_max_bsz, new_custom_sz);
-      CAML_GC_MESSAGE(PARAMS, "New custom minor size limit: %"
-                      ARCH_INTNAT_PRINTF_FORMAT "u%%\n",
-                      new_custom_sz);
-    }
-  }
-
-  /* Minor heap size comes last because it will trigger a minor collection
-     (thus invalidating [v]) and it can raise [Out_of_memory]. */
-  if (newminwsz != Caml_state->minor_heap_wsz) {
-    CAML_GC_MESSAGE(PARAMS, "New minor heap size: %"
-                    ARCH_INTNAT_PRINTF_FORMAT "uk words\n", newminwsz / 1024);
-  }
-
-  if (newminwsz > caml_minor_heap_max_wsz) {
-    CAML_GC_MESSAGE(PARAMS, "New minor heap max: %"
-                    ARCH_INTNAT_PRINTF_FORMAT "uk words\n", newminwsz / 1024);
-    caml_update_minor_heap_max(newminwsz);
-  }
-  CAMLassert(newminwsz <= caml_minor_heap_max_wsz);
-  if (newminwsz != Caml_state->minor_heap_wsz) {
-    /* FIXME: when (newminwsz > caml_minor_heap_max_wsz) and
-       (newminwsz != Caml_state->minor_heap_wsz) are both true,
-       the current domain reallocates its own minor heap twice. */
-    caml_set_minor_heap_size (newminwsz);
-  }
-
-  CAML_EV_END(EV_EXPLICIT_GC_SET);
-  return Val_unit;
-}
-
-CAMLprim value caml_gc_minor(value v)
-{
-  Caml_check_caml_state();
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_MINOR);
-  CAMLassert (v == Val_unit);
-  caml_minor_collection ();
-  caml_result result = caml_process_pending_actions_res();
-  CAML_EV_END(EV_EXPLICIT_GC_MINOR);
-  return caml_get_value_or_raise_async(result, "Gc.minor");
-}
-
-static caml_result gc_major_res(int compaction_mode)
-{
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_MAJOR);
-  CAML_GC_MESSAGE(MAJOR, "Major GC cycle requested\n");
-  caml_empty_minor_heaps_once();
-  caml_finish_major_cycle(compaction_mode);
-  caml_reset_major_pacing();
-  caml_result result = caml_process_pending_actions_res();
-  CAML_EV_END(EV_EXPLICIT_GC_MAJOR);
-  return result;
-}
-
-CAMLprim value caml_gc_major(value v)
-{
-  Caml_check_caml_state();
-  CAMLassert (v == Val_unit);
-  return caml_get_value_or_raise_async(gc_major_res(Compaction_auto),
-                                       "Gc.minor");
-}
-
-static caml_result gc_full_major_res(void)
-{
-  int i;
-  caml_result res;
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_FULL_MAJOR);
-  CAML_GC_MESSAGE(MAJOR, "Full Major GC requested\n");
-  /* In general, it can require up to 3 GC cycles for a
-     currently-unreachable object to be collected. */
-  for (i = 0; i < 3; i++) {
-    caml_finish_major_cycle(i == 2 ? Compaction_auto : Compaction_none);
-    caml_reset_major_pacing();
-    res = caml_process_pending_actions_res();
-    if (caml_result_is_exception(res)) break;
-  }
-  ++ Caml_state->stat_forced_major_collections;
-  CAML_EV_END(EV_EXPLICIT_GC_FULL_MAJOR);
-  return res;
-}
-
-CAMLprim value caml_gc_full_major(value v)
-{
-  Caml_check_caml_state();
-  CAMLassert (v == Val_unit);
-  return caml_get_value_or_raise_async(gc_full_major_res(), "Gc.full_major");
-}
-
-CAMLprim value caml_gc_major_slice (value v)
-{
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_MAJOR_SLICE);
-  CAMLassert (Is_long (v));
-  caml_major_collection_slice(Long_val(v));
-  caml_result result = caml_process_pending_actions_res();
-  CAML_EV_END(EV_EXPLICIT_GC_MAJOR_SLICE);
-  return caml_get_value_or_raise_async(result, "Gc.major_slice");
-}
-
-CAMLprim value caml_gc_compaction(value v)
-{
-  Caml_check_caml_state();
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_COMPACT);
-  CAMLassert (v == Val_unit);
-  caml_result res = Result_unit;
-  int i;
-  /* We do a full major before this compaction. See [caml_full_major_exn] for
-     why this needs three iterations. */
-  for (i = 0; i < 3; i++) {
-    caml_finish_major_cycle(i == 2 ? Compaction_forced : Compaction_none);
-    caml_reset_major_pacing();
-    res = caml_process_pending_actions_res();
-    if (caml_result_is_exception(res))
-      break;
-  }
-  ++ Caml_state->stat_forced_major_collections;
-  CAML_EV_END(EV_EXPLICIT_GC_COMPACT);
-  return caml_get_value_or_raise_async(res, "Gc.compact");
-}
-
-CAMLprim value caml_gc_stat(value v)
-{
-  caml_result result;
-  CAML_EV_BEGIN(EV_EXPLICIT_GC_STAT);
-  result = gc_full_major_res();
-  if (caml_result_is_exception(result)) goto out;
-  result = Result_value(caml_gc_quick_stat(Val_unit));
- out:
-  CAML_EV_END(EV_EXPLICIT_GC_STAT);
-  return caml_get_value_or_raise_async(result, "Gc.stat");
-}
-
-CAMLprim value caml_get_minor_free (value v)
-{
-  return Val_int
-    ((uintnat)Caml_state->young_ptr - (uintnat)Caml_state->young_start);
 }
 
 void caml_init_gc (void)
@@ -410,34 +115,6 @@ void caml_init_gc (void)
   caml_init_domains(caml_params->max_domains,
                     caml_params->init_minor_heap_wsz);
   caml_init_gc_stats(caml_params->max_domains);
-}
-
-/* FIXME After the startup_aux.c unification, move these functions there. */
-
-CAMLprim value caml_runtime_variant (value unit)
-{
-  CAMLassert (unit == Val_unit);
-#if defined (DEBUG)
-  return caml_copy_string ("d");
-#elif defined (CAML_INSTR)
-  return caml_copy_string ("i");
-#else
-  return caml_copy_string ("");
-#endif
-}
-
-/* Control runtime warnings */
-
-CAMLprim value caml_ml_enable_runtime_warnings(value vbool)
-{
-  caml_runtime_warnings = Bool_val(vbool);
-  return Val_unit;
-}
-
-CAMLprim value caml_ml_runtime_warnings_enabled(value unit)
-{
-  CAMLassert (unit == Val_unit);
-  return Val_bool(caml_runtime_warnings);
 }
 
 struct gc_tweak {
@@ -672,27 +349,4 @@ caml_result caml_gc_ramp_up(value callback, uintnat *out_suspended_words) {
 
 void caml_gc_ramp_down(uintnat ramp_up_words) {
   Caml_state->allocated_words_resumed += ramp_up_words;
-}
-
-CAMLprim value caml_ml_gc_ramp_up(value callback) {
-  CAMLparam1(callback);
-  CAMLlocal1(v);
-  uintnat deferred_words;
-  caml_result res = caml_gc_ramp_up(callback, &deferred_words);
-  if (caml_result_is_exception(res)) {
-    // We will re-raise the exception below; before that,
-    // we ramp_down to avoid discarding the deferred work.
-    caml_gc_ramp_down(deferred_words);
-  }
-  v = caml_get_value_or_raise(res);
-  CAMLreturn (caml_alloc_2(0, v, Val_long(deferred_words)));
-}
-
-CAMLprim value caml_ml_gc_ramp_down(value work) {
-  uintnat resumed_words = Long_val(work);
-  CAML_GC_MESSAGE(POLICY,
-    "GC ramp-down; resumed words: %"ARCH_INTNAT_PRINTF_FORMAT"u\n",
-    resumed_words);
-  caml_gc_ramp_down(resumed_words);
-  return Val_unit;
 }
